Backend/src/main/java/com/supermercado/market/Service/ClientService.java
package com.supermercado.market.Service;

import java.util.List;

import org.springframework.transaction.annotation.Transactional;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import com.supermercado.market.Domain.Client;
import com.supermercado.market.Repo.ClientRepository;


@Service
public class ClientService {

    @Autowired
    private ClientRepository clientRepository;

    public List<Client> getAllClients() {
        return clientRepository.findAll();
    }

    @Transactional
    public Client addClient(Client client) {
        return clientRepository.save(client);
    }

    public void deleteClient(Long id) {
        clientRepository.deleteById(id);
    }

    public Client updateClient(Long id, Client client) {
     
        Client existingClient = clientRepository.findById(id)
            .orElseThrow(() -> new RuntimeException("Cliente no encontrado con id: " + id));
        existingClient.setName(client.getName());
        existingClient.setRole(client.getRole());
        return clientRepository.save(existingClient);
    }

 
}